Fix output loop regression
